+++ src/api/src/main/java/api/auth/MuUserDetails.java
@@ -1,4 +1,4 @@
-package api.dto;
+package api.auth;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import io.micronaut.core.annotation.Introspected;
@@ -8,11 +8,11 @@ import java.util.Collections;
 import java.util.Optional;
 
 @Introspected
-public class UserDTO extends UserDetails {
+public class MuUserDetails extends UserDetails {
     public static final String ID = "id";
     private final String id;
 
-    public UserDTO(String id, String username) {
+    public MuUserDetails(String id, String username) {
         super(username, Collections.emptyList(), Collections.singletonMap(ID, id));
         this.id = id;
     }
